Remade the system used to get data, and created a preview data system
