--- Sources/SwiftVue/Preview/DistrictInfoPreviewData.swift
@@ -1,14 +1,0 @@
-//
-//  DistrictInfoPreviewData.swift
-//  
-//
-//  Created by Henry Pontzer on 8/23/23.
-//
-
-import Foundation
-let sampleDistrictInfo = DistrictInfo(
-    districtID: "",
-    name: "Synergy School District",
-    address: "East Fayetteville NC 28303",
-    url: "https://mystudentlogin.school.edu"
-)
