+++ Sources/SwiftVue/PreviewData/PreviewData.swift
@@ -1,0 +1,12 @@
+//
+//  File.swift
+//  
+//
+//  Created by Evan Kaneshige on 8/26/23.
+//
+
+import Foundation
+
+public struct PreviewData {
+    
+}
