+++ Sources/SwiftVue/StudentInfo/StudentInfoParser.swift
@@ -1,8 +1,8 @@
 //
-//  StudentInfo.swift
+//  StudentInfoParser.swift
 //  
 //
-//  Created by Evan Kaneshige on 12/7/22.
+//  Created by Evan Kaneshige on 8/26/23.
 //
 
 import Foundation
@@ -117,147 +117,5 @@ public class StudentInfoParser: NSObject, XMLParserDelegate {
             return .success(info)
         }
         
-    }
-}
-
-public struct StudentInfo: Hashable, Codable, Identifiable {
-    public var id: UUID
-    public var formattedName: String
-    public var permId: String
-    public var gender: String
-    public var grade: String
-    public var address: String
-    public var birthDate: String
-    public var email: String
-    public var phone: String
-    public var homeLanguage: String
-    public var currentSchool: String
-    public var homeRoomTeacher: String
-    public var homeRoomTeacherEmail: String
-    public var homeRoomNumber: String
-    public var counselorName: String
-    public var counselorEmail: String
-    public var base64Photo: String
-    public var emergencyContacts: [EmergencyContact]
-    public var physician: Physician
-    public var dentist: Dentist
-    public var userDefinedGroupBoxes: [UserDefinedGroupBox]
-    
-    public init(id: UUID = UUID(), formattedName: String, permId: String, gender: String, grade: String, address: String, birthDate: String, email: String, phone: String, homeLanguage: String, currentSchool: String, homeRoomTeacher: String, homeRoomTeacherEmail: String, homeRoomNumber: String, counselorName: String, counselorEmail: String, base64Photo: String, emergencyContacts: [EmergencyContact], physician: Physician, dentist: Dentist, userDefinedGroupBoxes: [UserDefinedGroupBox]) {
-        self.id = id
-        self.formattedName = formattedName
-        self.permId = permId
-        self.gender = gender
-        self.grade = grade
-        self.address = address
-        self.birthDate = birthDate
-        self.email = email
-        self.phone = phone
-        self.homeLanguage = homeLanguage
-        self.currentSchool = currentSchool
-        self.homeRoomTeacher = homeRoomTeacher
-        self.homeRoomTeacherEmail = homeRoomTeacherEmail
-        self.homeRoomNumber = homeRoomNumber
-        self.counselorName = counselorName
-        self.counselorEmail = counselorEmail
-        self.base64Photo = base64Photo
-        self.emergencyContacts = emergencyContacts
-        self.physician = physician
-        self.dentist = dentist
-        self.userDefinedGroupBoxes = userDefinedGroupBoxes
-    }
-    
-    public static func sample() -> StudentInfo {
-        return sampleStudentInfo
-    }
-}
-
-public struct EmergencyContact: Hashable, Codable, Identifiable {
-    public var id: UUID
-    public var name: String
-    public var relationship: String
-    public var homePhone: String
-    public var workPhone: String
-    public var otherPhone: String
-    public var mobilePhone: String
-    
-    public init(id: UUID = UUID(), name: String, relationship: String, homePhone: String, workPhone: String, otherPhone: String, mobilePhone: String) {
-        self.id = id
-        self.name = name
-        self.relationship = relationship
-        self.homePhone = homePhone
-        self.workPhone = workPhone
-        self.otherPhone = otherPhone
-        self.mobilePhone = mobilePhone
-    }
-    
-    public static func sample() -> EmergencyContact {
-        return sampleEmergencyContact
-    }
-}
-
-public struct Physician: Hashable, Codable, Identifiable {
-    public var id: UUID
-    public var name: String
-    public var hospital: String
-    public var phone: String
-    public var extn: String
-    
-    public init(id: UUID = UUID(), name: String, hospital: String, phone: String, extn: String) {
-        self.id = id
-        self.name = name
-        self.hospital = hospital
-        self.phone = phone
-        self.extn = extn
-    }
-    
-    public static func sample() -> Physician {
-        return samplePhysician
-    }
-}
-
-public struct Dentist: Hashable, Codable, Identifiable {
-    public var id: UUID
-    public var name: String
-    public var office: String
-    public var phone: String
-    public var extn: String
-    
-    public init(id: UUID = UUID(), name: String, office: String, phone: String, extn: String) {
-        self.id = id
-        self.name = name
-        self.office = office
-        self.phone = phone
-        self.extn = extn
-    }
-    
-    public static func sample() -> Dentist {
-        return sampleDentist
-    }
-}
-
-public struct UserDefinedGroupBox: Hashable, Codable, Identifiable {
-    public var id: UUID
-    public var userDefinedItems: [UserDefinedItem]
-    
-    public init(id: UUID = UUID(), userDefinedItems: [UserDefinedItem]) {
-        self.id = id
-        self.userDefinedItems = userDefinedItems
-    }
-}
-
-public struct UserDefinedItem: Hashable, Codable, Identifiable {
-    public var id: UUID
-    public var itemLabel: String
-    public var value: String
-    
-    public init(id: UUID = UUID(), itemLabel: String, value: String) {
-        self.id = id
-        self.itemLabel = itemLabel
-        self.value = value
-    }
-    
-    public static func sample() -> UserDefinedItem {
-        return sampleUserDefinedItem
     }
 }
